requirements.txt
# Discord
discord.py-self

# Formatter
black
isort

# LLM
litellm

# Utils
setuptools
requests
colorama
python-dotenv

# Network
requests
beautifulsoup4
playwright

# Downloader
downloadm3u8